修改创建订单的逻辑，在收到回调通知后弹出电池，更新订单
