+++ zprovider/src/main/java/com/ycb/zprovider/controller/CreditUploadController.java
@@ -30,8 +30,6 @@ import java.util.Date;
 public class CreditUploadController {
     public static final Logger logger = LoggerFactory.getLogger(CreditUploadController.class);
     //初始化alipayClient用到的参数:支付宝网关
-    @Value("${serverUrl}")
-    private String serverUrl;
     //初始化alipayClient用到的参数:该appId必须设为开发者自己的生活号id
     @Value("${appId}")
     private String appId;
@@ -76,7 +74,8 @@ public class CreditUploadController {
         //注：商户维度下，类目Code（categoryCode）+实体编号（entity_code）唯一，
         // 一个商户下相同类目code+实体编号多次调用，将按照上传时间（upload_time）更新，
         // 更新规则取最新的upload_time快照数据
-        String entityCode = shopStation.getId().toString();
+//        String entityCode = shopStation.getId().toString();
+        String entityCode = "10";
         //地址位置经度，取值范围：经度-180~180，中国地区经度范围：73.66~135.05，示例：83.66
         String longitude = shopStation.getLongitude();
         //地址位置纬度，取值范围：纬度-90~90，中国地区经度范围：纬度3.86~53.55	，示例：5.87
